refactor(AddressBar, Browser, browserReducer, browser.js): remove unused code and fix newTabOpenedTabs function
The AddressBar component had a typo in the PropTypes definition for the tabCount property. The apply method was removed.

The Browser component had an unused function called newTabOpenedTabs that was removed. The newTabRight function was fixed to create a new tab on the right side of the current tab.

The browserReducer had a bug where the activeIndex was being set to the wrong value when creating a new tab. This was fixed.

The browser.js utility file had a bug where the activeTabId was not being removed from AsyncStorage when the activeTabRef was null. This was fixed by using the removeItem method instead of setting the value to null.